Track click centering & zooming, plus related fixes.
Fixes 'center-on-click' on ideogram.  Uses a new CenterProperty w/event
listener to signal centering clicks from RefFrame to ContentPane.  The
actual CenterProperty value is not yet used.
Implements 'center-on-click' and 'double-click-to-zoom' on Tracks.
Draws a border around FeatureTrack as a temporary measure to show
position during UI development.
Minor clean-up in CytobandPane.
Implements clearing of Tracks on Genome change.