Fixes the track sizing feature
Changing the track size now adjusts the TrackScrollPane instead of the
Track itself (which is sized by content).
Also added a secret UI-dev control to randomly change the track content
size so we can more easily see the interaction with the TrackScrollPane
and the effects as the size of each change.  Hold down the Control key
while double-clicking on the track to use it.