Implements a ProgressBarDialog
This is intended to be a generic component, though for now it's only
used for loading genomes.  We might find it needs an update eventually.
It only gives an indeterminate indicator; eventually we may need to
extend it to show e.g. percent of progress.
The old indicator still needs to be cleaned up.